refactoring - extract table composition code from TableExtraction to TableComposer
