+++ TextExtraction/lib/table-composition/Lines.h
