+++ TextExtraction/lib/table-composition/TableComposer.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "Table.h"
+#include "Lines.h"
+
+#include "../text-parsing/ParsedTextPlacement.h"
+
+class TableComposer {
+    public:
+        TableComposer();
+        virtual ~TableComposer();
+
+        TableList ComposeTables(const Lines& inLines, const ParsedTextPlacementList& inTextPlacements);
+};
